progress: replace setitimer() with alarm()
setitimer() was marked obsolescent by IEEE Std 1003.1-2017 and is
no longer available in the latest edition.

While other alternatives of high fidelity are available, its use
in progress only requires whole second intervals, so it can be
cleanly replaced by a simpler alarm(); do that.

MinGW provided its own version of setitimer() so add an equivalent
implementation for alarm(), but do the minimum changes required in
the timer thread to adapt to the new interface.

Remove the compatibility layer for setitimer() and its related
struct that are no longer needed, and adjust the build system.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>