combine-diff: don't override recursive flag
Function diff_tree_combined() copies the 'struct diff_options' from the
input 'struct rev_info' to override some flags. One flag was
'recursive', which is set to 1. This has been the case since the
inception of this function in af3feefa1d (diff-tree -c: show a merge
commit a bit more sensibly., 2006-01-24). From that commit there's no
clear indication why recursive is overridden. But it seems to be
incorrect.

Looking at a test case in t4013. The expected test results are in
t/t4013/. And take 'diff.noellipses-diff-tree_-c_--abbrev_master' as an
example:

   $ git diff-tree -c --abbrev master
   59d314ad6f356dd08601a4cd5e530381da3e3c64
   ::100644 100644 100644 cead32e 7289e35 992913c MM	dir/sub
   ::100644 100644 100644 b414108 f4615da 10a8a9f MM	file0
   $

Here we see the expected results include 'dir/sub', while argument
'-r' (for recursive) was not given. So we should expect 'dir' instead.

Stop overriding the 'recursive' flag in diff_tree_combined(). This
changes the expected results for the example above to be:

    $ git diff-tree -c --abbrev master
    59d314ad6f356dd08601a4cd5e530381da3e3c64
    ::040000 040000 040000 65f5c9d f977ed4 0564e02 MM	dir
    ::100644 100644 100644 b414108 f4615da 10a8a9f MM	file0
    $

This also fixes the recently added subcommand git-last-modified(1),
which in some cases (when merges are involved) exited with the error:

    BUG: paths remaining beyond boundary in last-modified

The last-modified machinery uses a hashmap for all the files it wants to
get the last-modified commit for. Through log_tree_commit() the callback
mark_path() is called. Here the incoming path is looked up in the
hashmap, but because the diff-tree machinery internally switched to
recursive (even if the last-modified machinery wasn't) the entry for
'dir' was never expelled from the hashmap, and the BUG() statement was
hit.

Because internally diff-tree no longer runs recursive, this results in a
nice speedup when running `git last-modified` on git.git:

    Benchmark 1: next
      Time (mean ± σ):      6.068 s ±  0.038 s    [User: 5.268 s, System: 0.187 s]
      Range (min … max):    6.007 s …  6.123 s    10 runs

    Benchmark 2: HEAD
      Time (mean ± σ):      3.057 s ±  0.009 s    [User: 2.996 s, System: 0.055 s]
      Range (min … max):    3.047 s …  3.070 s    10 runs

    Summary
      HEAD ran
        1.98 ± 0.01 times faster than next

Besides that, set the flag 'recursive' in submodule.c to 1 because it
calls diff_tree_combined_merge() which no longer enables the flag
itself.

Signed-off-by: Toon Claes <toon@iotcl.com>